requirements.txt
streamlit
moviepy
openai
SpeechRecognition
pydub
requests
pyttsx3
python-dotenv